+++ src/Sulu/Bundle/AdminBundle/Navigation/ContentNavigationItem.php
@@ -83,6 +83,13 @@ class ContentNavigationItem
      */
     private $position;
 
+    /**
+     * Defines conditions whether the tab will be displayed or not.
+     *
+     * @var DisplayCondition[]
+     */
+    private $displayConditions;
+
     public function __construct($name)
     {
         $this->name = $name;
@@ -232,5 +239,21 @@ class ContentNavigationItem
     public function setPosition($position)
     {
         $this->position = $position;
+    }
+
+    /**
+     * @return DisplayCondition[]
+     */
+    public function getDisplayConditions()
+    {
+        return $this->displayConditions;
+    }
+
+    /**
+     * @param DisplayCondition[] $displayConditions
+     */
+    public function setDisplayConditions(array $displayConditions)
+    {
+        $this->displayConditions = $displayConditions;
     }
 }