+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/collections/edit/main.js
@@ -52,7 +52,10 @@ define([
 
                 noBack: true,
                 tabs: {
-                    url: '/admin/content-navigations?alias=media'
+                    url: '/admin/content-navigations?alias=media',
+                    componentOptions: {
+                        values: this.data
+                    }
                 },
                 toolbar: {
                     buttons: buttons,
@@ -165,6 +168,7 @@ define([
             this.sandbox.on('sulu.medias.collection.saved', function(id, collection) {
                 if (!collection.locale || collection.locale === UserSettingsManager.getMediaLocale()) {
                     this.data = collection;
+                    this.sandbox.emit('sulu.header.saved', this.data);
                 }
             }.bind(this));